Time Series Forecasting for Taxi Demand
Project: Taxi Order Prediction

Taxi company has collected historical data on taxi orders at airports. To attract more drivers during peak hours, it is necessary to forecast the number of taxi orders for the next hour. The goal of this project is to build a machine learning model for such a prediction.

Success criterion: the RMSE metric on the test set should be **no more than 48**.